backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.7
django-cors-headers==4.3.1
django-filter==23.3
pytest==7.4.3
pytest-django==4.7.0
django-environ==0.11.2
